# src/spec/api.yaml
openapi: 3.0.3
info:
  title: Actor API
  description: API каталога актёров (без эндпоинтов входа/выхода админа)
  version: 1.0.0

servers:
  - url: /v1
    description: API v1

paths:
  # Создать вуз
  /universities:
    post:
      summary: Создать вуз
      operationId: v1UniversityCreatePost
      tags: [universities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/UniversityCreate'
      responses:
        '201':
          description: Результат создания (id + status ok/failed + errorCode)
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/UniversityCreateResponse'

  # 1) Создать актёра
  /actors:
    post:
      summary: Создать актёра
      operationId: v1ActorCreatePost
      tags: [actors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/ActorCreate'
      responses:
        '201':
          description: Результат создания (id + status ok/failed + errorCode)
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ActorCreateResponse'

    # 4) Получить актёров по фильтрам
    get:
      summary: Получить актёров по фильтрам
      operationId: v1ActorsGet
      tags: [actors]
      parameters:
        - name: gender
          in: query
          schema:
            $ref: './schemas.yaml#/components/schemas/Gender'
        - name: ageFrom
          in: query
          schema:
            type: integer
          description: Возраст от (лет)
        - name: ageTo
          in: query
          schema:
            type: integer
        - name: weightMin
          in: query
          schema:
            type: integer
        - name: weightMax
          in: query
          schema:
            type: integer
        - name: activityYearFrom
          in: query
          schema:
            type: integer
          description: Годы активности от
        - name: activityYearTo
          in: query
          schema:
            type: integer
        - name: universityId
          in: query
          schema:
            $ref: './schemas.yaml#/components/schemas/MongoId'
          description: ID вуза
        - name: theatre
          in: query
          schema:
            type: string
        - name: title
          in: query
          schema:
            $ref: './schemas.yaml#/components/schemas/Title'
        - name: hairColor
          in: query
          schema:
            type: string
        - name: eyeColor
          in: query
          schema:
            type: string
        - name: genres
          in: query
          schema:
            type: array
            items:
              type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Список актёров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './schemas.yaml#/components/schemas/Actor'

  # 2) Получить актёра по id
  /actors/{id}:
    get:
      summary: Получить актёра по ID
      operationId: v1ActorByIdGet
      tags: [actors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: './schemas.yaml#/components/schemas/MongoId'
      responses:
        '200':
          description: Актёр
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Actor'
        '404':
          description: Не найден

  # 3) Загрузить медиа (Multipart/Form-Data, тип фото/видео, GridFS)
  /actors/{id}/media:
    post:
      summary: Загрузить медиа актёра (фото или видео в GridFS)
      operationId: v1ActorMediaUploadPost
      tags: [media]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: './schemas.yaml#/components/schemas/MongoId'
          description: ID актёра
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, type]
              properties:
                file:
                  type: string
                  format: binary
                type:
                  $ref: './schemas.yaml#/components/schemas/ActorMediaType'
                caption:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Результат загрузки (mediaId + status ok/failed + errorCode)
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/MediaUploadResponse'
        '400':
          description: Ошибка запроса
        '404':
          description: Актёр не найден

  # 7) Получить медиа по id (бинарный контент)
  /actors/{actorId}/media/{mediaId}:
    get:
      summary: Получить медиафайл по id (бинарный контент из GridFS)
      operationId: v1MediaByIdGet
      tags: [media]
      parameters:
        - name: actorId
          in: path
          required: true
          schema:
            $ref: './schemas.yaml#/components/schemas/MongoId'
        - name: mediaId
          in: path
          required: true
          schema:
            $ref: './schemas.yaml#/components/schemas/MongoId'
          description: ID элемента в photos или videos
      responses:
        '200':
          description: Файл (image/* или video/*)
          content:
            image/*:
              schema:
                type: string
                format: binary
            video/*:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Медиа не найдено